requirements.txt
crewai
langchain-openai
python-dotenv
fastapi
uvicorn
tavily-python
firecrawl-py
pinecone-client
numpy
pyyaml
markdown
requests